clutter/stage: Add infrastructure to track devices and their actors
For the input thread to get to its full potential we need to move the
tracking of actors associated with input devices out of
ClutterInputDevice. We're going to move it into ClutterStage instead.

So start that by adding the infrastructure to ClutterStage to keep track
of those things. It consists of two hashtables which associate devices
and touch sequences with actors, those hashtables get updated using
clutter_stage_associate_actor_device() and can be queried by calling
clutter_stage_get_device_actor(), which will replace
clutter_input_device_get_actor().