clutter: Add function to update inputdevice actors to clutter-main
We're moving keeping track of inputdevices and their associated actors
out of ClutterInputDevice, this commit basically adds replacements for
clutter_input_device_update() and clutter_input_device_set_actor() to
clutter-main and shuffles the internals of those functions around a bit
for clarity.

clutter_update_device_actor() is made public because we need to update
the actor of inputdevices not only from ClutterStage (when repicking
after relayout), but also from MetaWaylandPointer.