launcher: Ignore session active changes while logind is gone
When logind disconnects from the bus, property change notifications will
be emitted and we'll receive a default False value for "active".

However, when this happens we should assume that the active state
remains the same and will not change until logind reappears on the bus.
As such, check whether there is a name owner and only continue with
syncing if logind is running.

See https://github.com/systemd/systemd/issues/17308

https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/1556